# contracts/openapi-v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Quicker.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/numbering/series": {
      "get": {
        "tags": [
          "Numbering"
        ],
        "operationId": "getNumberingSeries",
        "parameters": [
          {
            "name": "documentType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "companyId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SeriesSummary"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Numbering"
        ],
        "summary": "Create a series: template with {company} {branch} {yy} {yyyy} {mm} {fy} and one {seq[:N]}, gapless or not, reset never/yearly/monthly",
        "operationId": "postNumberingSeries",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveSeriesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesSummary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/numbering/series/{seriesId}": {
      "get": {
        "tags": [
          "Numbering"
        ],
        "operationId": "getNumberingSeriesBySeriesId",
        "parameters": [
          {
            "name": "seriesId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesSummary"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Numbering"
        ],
        "operationId": "putNumberingSeriesBySeriesId",
        "parameters": [
          {
            "name": "seriesId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveSeriesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesSummary"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/numbering/preview": {
      "post": {
        "tags": [
          "Numbering"
        ],
        "summary": "Which series a document would use and what its next number would look like (no allocation)",
        "operationId": "postNumberingPreview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AllocateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NumberPreview"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/numbering/allocate": {
      "post": {
        "tags": [
          "Numbering"
        ],
        "summary": "Allocate a number for a document managed outside Quicker (integrations); documents inside Quicker allocate in their posting transaction",
        "operationId": "postNumberingAllocate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AllocateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllocatedNumber"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/numbering/drafts/{documentId}": {
      "get": {
        "tags": [
          "Numbering"
        ],
        "operationId": "getNumberingDraftsByDocumentId",
        "parameters": [
          {
            "name": "documentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/integration/webhooks": {
      "get": {
        "tags": [
          "Integration"
        ],
        "operationId": "getIntegrationWebhooks",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WebhookSummary"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Integration"
        ],
        "summary": "Create a subscription; the response carries the signing secret once",
        "operationId": "postIntegrationWebhooks",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveWebhookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookCreated"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/integration/webhooks/{webhookId}": {
      "get": {
        "tags": [
          "Integration"
        ],
        "operationId": "getIntegrationWebhooksByWebhookId",
        "parameters": [
          {
            "name": "webhookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookSummary"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Integration"
        ],
        "operationId": "putIntegrationWebhooksByWebhookId",
        "parameters": [
          {
            "name": "webhookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveWebhookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookSummary"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Integration"
        ],
        "operationId": "deleteIntegrationWebhooksByWebhookId",
        "parameters": [
          {
            "name": "webhookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/v1/integration/webhooks/{webhookId}/rotate-secret": {
      "post": {
        "tags": [
          "Integration"
        ],
        "operationId": "postIntegrationWebhooksByWebhookIdRotateSecret",
        "parameters": [
          {
            "name": "webhookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookCreated"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/integration/webhooks/{webhookId}/test": {
      "post": {
        "tags": [
          "Integration"
        ],
        "summary": "Queue a webhook.test delivery so the receiver can check URL, signature and parsing",
        "operationId": "postIntegrationWebhooksByWebhookIdTest",
        "parameters": [
          {
            "name": "webhookId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeliverySummary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AllocatedNumber": {
        "required": [
          "seriesId",
          "seriesCode",
          "periodKey",
          "number",
          "text",
          "gapless"
        ],
        "type": "object",
        "properties": {
          "seriesId": {
            "type": "string",
            "format": "uuid"
          },
          "seriesCode": {
            "type": "string"
          },
          "periodKey": {
            "type": "string"
          },
          "number": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "text": {
            "type": "string"
          },
          "gapless": {
            "type": "boolean"
          }
        },
        "description": "An allocated number: the series that issued it, its sequence within the reset period, and the rendered text."
      },
      "AllocateRequest": {
        "required": [
          "documentType",
          "companyId",
          "date",
          "documentId"
        ],
        "type": "object",
        "properties": {
          "documentType": {
            "type": "string"
          },
          "companyId": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "documentId": {
            "type": "string",
            "format": "uuid"
          },
          "branchId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "seriesId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
...
      "CounterSummary": {
        "required": [
          "periodKey",
          "nextNumber",
          "allocated",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "periodKey": {
            "type": "string"
          },
          "nextNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "allocated": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "DeliverySummary": {
        "required": [
          "id",
          "subscriptionId",
          "eventId",
          "eventType",
          "payload",
          "attempt",
          "status",
          "responseStatus",
          "responseExcerpt",
          "lastError",
          "jobId",
          "createdAt",
          "attemptedAt",
          "nextAttemptAt",
          "deliveredAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "subscriptionId": {
            "type": "string",
            "format": "uuid"
          },
          "eventId": {
            "type": "string",
            "format": "uuid"
          },
          "eventType": {
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/JsonElement"
          },
          "attempt": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "status": {
            "type": "string"
          },
          "responseStatus": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "responseExcerpt": {
            "type": [
              "null",
              "string"
            ]
          },
          "lastError": {
            "type": [
              "null",
              "string"
            ]
          },
          "jobId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "attemptedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "nextAttemptAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "deliveredAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "JsonElement": { },
...
      "NumberPreview": {
        "required": [
          "seriesId",
          "seriesCode",
          "periodKey",
          "nextNumber",
          "text",
          "gapless"
        ],
        "type": "object",
        "properties": {
          "seriesId": {
            "type": "string",
            "format": "uuid"
          },
          "seriesCode": {
            "type": "string"
          },
          "periodKey": {
            "type": "string"
          },
          "nextNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "text": {
            "type": "string"
          },
          "gapless": {
            "type": "boolean"
          }
        },
        "description": "The series a request would use and the text its next number would have (no lock, no allocation)."
      },
...
      "SaveSeriesRequest": {
        "required": [
          "code",
          "documentType",
          "companyId",
          "template"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "documentType": {
            "type": "string"
          },
          "companyId": {
            "type": "string",
            "format": "uuid"
          },
          "template": {
            "type": "string"
          },
          "branchId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "fiscalYearId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "startNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64",
            "default": 1
          },
          "gapless": {
            "type": "boolean",
            "default": true
          },
          "resetPolicy": {
            "type": "string",
            "default": "never"
          },
          "validFrom": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "validTo": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "isDefault": {
            "type": "boolean",
            "default": false
          },
          "isActive": {
            "type": "boolean",
            "default": true
          }
        }
      },
...
      "SaveWebhookRequest": {
        "required": [
          "name",
          "url"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "eventTypes": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "filters": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "type": "string"
            }
          },
          "active": {
            "type": "boolean",
            "default": true
          }
        }
      },
...
      "SeriesSummary": {
        "required": [
          "id",
          "code",
          "documentType",
          "companyId",
          "branchId",
          "fiscalYearId",
          "template",
          "startNumber",
          "gapless",
          "resetPolicy",
          "validFrom",
          "validTo",
          "isDefault",
          "isActive",
          "counters"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string"
          },
          "documentType": {
            "type": "string"
          },
          "companyId": {
            "type": "string",
            "format": "uuid"
          },
          "branchId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "fiscalYearId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "template": {
            "type": "string"
          },
          "startNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "gapless": {
            "type": "boolean"
          },
          "resetPolicy": {
            "type": "string"
          },
          "validFrom": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "validTo": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "isDefault": {
            "type": "boolean"
          },
          "isActive": {
            "type": "boolean"
          },
          "counters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CounterSummary"
            }
          }
        }
      },
...
      "WebhookCreated": {
        "required": [
          "subscription",
          "secret"
        ],
        "type": "object",
        "properties": {
          "subscription": {
            "$ref": "#/components/schemas/WebhookSummary"
          },
          "secret": {
            "type": "string"
          }
        },
        "description": "Returned once, at creation and rotation: the secret is stored encrypted and never shown again."
      },
      "WebhookSummary": {
        "required": [
          "id",
          "name",
          "url",
          "eventTypes",
          "filters",
          "active",
          "createdBy",
          "createdAt",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "eventTypes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "filters": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "active": {
            "type": "boolean"
          },
          "createdBy": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
}